changed dolt schema to be a command group
