+++ bats/diff.bats
@@ -109,7 +109,7 @@ teardown() {
     dolt table import -c -s=`batshelper employees-sch.json` employees `batshelper employees-tbl.json`
     dolt add employees
     dolt commit -m "Added employees table with data"
-    dolt schema --add-column employees city string
+    dolt schema add-column employees city string
     dolt table put-row employees id:3 "first name":taylor "last name":bantle title:"software engineer" "start date":"" "end date":"" city:"Santa Monica"
     run dolt diff --summary
     [ "$status" -eq 0 ]
