+++ bats/multiple-tables.bats
@@ -22,7 +22,7 @@ teardown() {
       [[ "$output" =~ "test1" ]] || false
       [[ "$output" =~ "test2" ]] || false
       [ "${#lines[@]}" -eq 3 ]
-      run dolt schema
+      run dolt schema show
       [ "$status" -eq 0 ]
       [[ "$output" =~ "test1 @ working" ]] || false
       [[ "$output" =~ "test2 @ working" ]] || false
